+++ src/master/pool.h
@@ -15,6 +15,6 @@ struct PoolHost *pool_host_lookup(union WireLink *link);
 
 struct PoolHost *pool_handle_new(uint32_t *room_out, bool random);
 struct PoolHost *pool_handle_new_named(uint32_t *room_out, ServerCode code);
-void pool_handle_free(struct PoolHost *host, uint16_t room);
+void pool_handle_free(struct PoolHost *host, uint32_t room);
 ServerCode pool_handle_code(struct PoolHost *host, uint32_t room);
 struct PoolHost *pool_handle_lookup(uint32_t *room_out, ServerCode code);
